Use empty user's name and email if git doesn't know them
For the new git installations the "user.name" and "user.email" settings
may not be configured. As the result, if a user tries to use "git
elegant configure ..." directly or as part of another command ("init"
etc.), it leads to a breaking of "configure" execution since default
setting can't be read.

As a solution, an empty value will be used by default in the case of the
"user.name" and "user.email" aren't configured on a global level.

#110