Propagate user arguments to libexec scripts
`bin/git-elegant` is the main entry script which has to pass all
arguments to `libexec/git-elegant` script, otherwise, a help message
will be printed and there is no chance to call any other command -
in other words, there is a bug. Since this change, all arguments will be
propagated to right place and "Elegant git" should work as expected.

Testing approach is based on invocation an unknown command after the
installation. This initiates the real call to "Elegant git" logic. So,
the arguments propagation logic is tested. Previous solution (invoke
`help` command) is not a good choice as help message can be shown in
many cases (untrustworthy results of testing).