Describe contribution process
As a potential contributor starts from README.md file, it contains
intical information about the contribution process. Then, it redirects
to CONTRIBUTING.md which provides a lot of details. Finally, the most
useful development info is presented in the README.md as it should be
quickly accesible.

Also, a pull request template is added. It should simplify the experience
of working on a pull request fro both contributor and reviewer as it was
buit based on the comments for previous pull requests.