Add help messages for each Elegant Git command
`-h`, `--help`, `help` are allowed options to get a help message for any
command. The `libexec/git-elegant` is responsible for building and
printing help. The commands scripts are responsible for providing the
messages. It is described on "Architecture" section of README.md.

Also, there are new tests which allow checking of the availability of the
help messages.

#118