Don't use `obtain-work` as a checkout mechanism in `start-work`
Since release 0.5.0, `obtain-work` works only with remote
braches. But we try to use it as a checkout mechanism for a local
branch in `start-work` which breaks its execution.

In order to resolve this issue, the following is made:
1. The checkout procedure of `start-work` is replaced with alternative
`git` commands.
2. Now the `addons-fake.bash` supports redirection of incomming
request to original command if there is no mocked command. This allows
having partially-mocked executions for testing (previously, either usage
of real git or faked command was allowed).

Also, one additional rule is added to the "Writing tests" section. It
was added as faking a `git-elegant` commands lead to such issues. In
the future, any newly-created or updated test(s) should follow it.