Make stashing verbose for `start-work` command
The introduced `pipe` plugin gives ability to execute a command through
a wrapper function which runs pre- and/or post-conditions. For now, there
is an ability to make automatic stash if there are some uncommitted
changes, and apply them back after the wrapped command is executed.

Use `git rev-parse --abbrev-ref HEAD` to display a name of the current
branch. As the result:
1. it will be executed faster
2. removes `*` prefix from the branch name

Use `$@` always in double-quotes. This makes correct parameters
propagation for all arguments.