Automate development chore using single entry script
`workflows` script provides capabilities on managing development
routine. It is a single entry point to run any developers command
and is used in CI setup.
In the future, if there is an action which simplifies development, it
must be added to the script.

Any other subscripts are stores within `.workflows` directory. This
cleans root directory and keeps non-production code separated.