Use `/tmp` directory as a default one during testing
Since `/tmp` is used as a working directory for BATS tests, there is no
chance that any unexpected modifications will happen within mounted
project directory.

Also, the `/tmp` does not have an initialized git repository. This is
good as if we need a real repository, we have to use addon for it.