Use correct git URL for installation
The URL wasn't updated during repository migration from a personal account
to the Bees Hive organization. As I still have a fork, installation works,
but it installs the obsolete version.

Now, the installation will use the correct git  repository and always use
the latest version of Elegant Git.

is not always up to date.