Display installed version of Elegant Git
The version will be automatically read from the git history and placed
into a file. Elegant Git will read it from the file when a user requests
a version.

Also, `--depth` is set 50 in order to make `git describe` always
working. We usually make small releases and the number should guarantee
stable execution.

#198