Define workflows for the project
Before stashing changes, the documentation will be generated. After
making a commit, the full CI flow will be triggered.

After a release, either a pull request for brew formulae will be
submitted or the instructions will be printed.

Also, the `version` file is removed before running a CI task in order to
prevent a failure of the version test. For unit tests, the file is not
expected. But after the installation, the file will be in the project's
directory.