Skip workflows if needed
Sometimes, a user may want to ignore defined workflows and just execute
a pure command. And `--no-workflows` option is designed to implement
this behavior. This option can be put prior or after the command. This
gives more flexibility in usage. However, `git-elegan --no-workflows
<command>` is preffered as it encourages to split Elegant Git options
and command arguments. Bash completion is updated appropriately.

Also, two improvements:
1. The `version` file will be removed prior to actual test execution.
This exposes the mentioned behavior for any place where
`.workflows/bats-pipeline.bash` is used.
2. Remove `-v` option (was equal to `--version`) as, in Git, `-v` stands
for `--verbose`.

#192