Introduce `acquire-git` command
The command allows applying the required configuration to a Git
installation. Since there are some interceptionsbetween the
`acquire-git` and `acquire-repository`, there are some changes in the
overall configuration workflow. These changes are desribed on the
documentation.

Also, to prevent unexpected tests failures, the `~/.gitconfig` is
removed after each test where a real Git repository is used.

#177