Simplify the configuration process
The configuration will start from `acquire-git` command and guide a user
through the process interactively. If standards are configured, they
will be printed to CLI instead of reconfiguring.

During the installation, `acquire-git` will be executed instead of
providing the instructions. This allows updating settings for the new
versions automatically.

Now, the `read` addon supports partial mocking of the answers correctly.
This works because a replay variable is always set with either provided
value or empty value.

#177