Add the ability to polish current work
When you are working in a separate branch, you may create several
commits. And once the work is done, you may want to change the commits
history somehow. `git rebase --interactive` can help with this. But it
requires an exact number of commits prior to running the command. So,
you have to provide it. And usually, it's done by reviewing `git log`
and counting of necessary commits.

`polish-work` command aims to solve this in a single command run. And
even, if there are pauses in the rebase process (like conflicts or
purposeful stops), the command can continue the started rebase process.