Use `vi` as the default editor during testing and debugging
All testing and debugging activities are performed within a Docker
container. And sometimes, the real editor is required to make some
experiments. As the dummy editor was used, some commands were failed as
it does not work. For instance, when an interactive rebase is executed.

As the solution, the `vi` editor now is used the default editor for git.
It is available in our alpine-based Docker image, that's why the CI will
work without changes.