Apply only created stashes automatically
There are cases when Git says about modification, although they are
absent in a working tree. And Elegant Git, as it sees the changes, will
try to stash them. But the stash won't be created (Git just says: "No
local changes to save"), and, as the exit code is 0, the next command
will be executed. At this point, Elegant Git will assume, that created
stash (which is false) must be uncovered after commands execution. And
it will apply the last available stash instead of doing nothing. That's
why a check is added that sees if there is a real created stash
that must be uncovered.