Don't ask issue number in the pull request template
By the contribution guide, a contributor should specify the issue number
in a commit message if possible. That's why there is on sense to ask
about it again in a pull request. Additionally, if there is only one
commit, its message will be inserted prior to template message while
creating a new pull request. So, this simple change should speed up
the creation of the pull requests.