Remove useless branches in a repository
The purpose of `clear-local` command is to remove branches that have
unexisting upstreams. And it starts asking for deleting branches if they
aren't merged. But this behavior is overengineering and, usually, all
these branches have to be removed.

As the solution, the `prune-repository` command is introduced which
makes safe automatic cleanup of all useless branches. Also, it doesn't
try to save your working state by using pipes. This is because that if
`prune-repository` is executed, then the repository should not have
uncommitted changes, otherwise, it will fail.