Apply preserves stash after rerun of failed command
There are uncommitted changes in the repository. If a command, which uses
a stash pipe, is executed, a stash will be created. But the command can
have the following execution result: passed or failed. If it is passed,
then the stash will be applied back. But if it fails, the stash will remain.
And now, if you rerun the same command and it will be successful, the
stash will be applied back. In other words, the state of stashing is
preserved for the command and will be applied back once possible.