Apply preserves branch after rerun of failed command
There is a working branch. If a command, which uses a stash pipe (like
release-work), is executed, the branch will be saved prior to execution
and restored after. But the command can have the following execution
result: passed or failed. If it is passed, then the current working
tree will be the working branch. But if it fails, the checkout won't
be performed. And now, if you rerun the same command and it will be
successful, the checkout of  the working branch will be performed.