Support active rebase process during `accept-work` execution
It happens that there are conflicts during the rebase of the latest
upstream into current branch for acceptance. And once they are solved,
the `accept-work` execution one more time allows continuing current
rebase and accepting work. But if a rebase is initiated not by
`accept-work` command, the execution will fail with appropriate message.

#146