Support workflows from repository subtrees
During development, the non-root repository directory can be used as
well as a root also. But, regardless of the position in a repository
directories tree, the workflows have toi be executed in the same way.
And it is achieved by updating a workflow file path with the path of
the top-level directory relative to the current directory.

Also, git-elegant.bats uses the real repository for testing since it's
required for workflows testing.

The tests for `clone-repository` and `show-commands` are still support
execution without the real Git repository.