Make Elegant Git working with unavailable remote
A user can do some work in a repository even if there is no access to a
remote. For instance, there is no internet available. In this case,
Elegant Git should be still operable. That's why `prune-repository` and
`start-work` ignore the availability of the remote. All other commands
either can't achieve their purpose without alive remote or don't use it
at all.

#116