Use the latest changes of the `master` branch during pruning repository
There are 2 states of the repository:

1. There is a remote configured. This means that the `master` branchi
has an upstream (usually `origin/master`). And after `git fetch --all`,
the `git rebase` has to be executed in order to rebase (merge) fetched
changes to сurrent local branch. Having the latest version is critical
as the `master` branch is used to seeking for useless local branches.

2. There is no remote configured. This means that the `master` branch
is up to date and seeking for useless local branches should be
performed.