Prevent "not a git repository" error during initiating a repository
Prior to the execution of any command, Elegant Git seeks the workflow
files. And in order to solve them from any position in a working tree,
it uses Git command for resolving a path to the root of the repository.
In the case, if `init-repository` or `clone-repository` is executed,
Git will raise an error: "not a git repository (or any of the parent
directories): .git." This happens as there is no `.git` directory at the
moment of the command execution.

Now, for both commands (`init-repository` and `clone-repository`), Git
command for the path evaluation won't be executed. This does not affect
the logic of the workflow execution but prevents the error.