Support untagged repository during the release process
If the release process starts for a repository without tags, all commits
have to be included in the release. This is achieved by introducing a
`diapason` variable for specifying a revisions range in both
`release-work` and `show-release-notes` commands. The variable is set to
head if there are no tags in the repository. And, since the
`show-release-notes` is always called with arguments within
`release-work`, a `from-reference` argument has `all-commits` value that
allows to recognize untagged repository and displays a release log that
includes all commits.

If the repository has at least one tag, the release process works
without changes.

#240