Start work from the alternative branch
Sometimes a new work should be started from the non-master branch. And
this is implemented by adding a second optional argument to
`start-work` command. Also, internal logic has an enhancement that
allows pulling only for branches that have configured remote tracking.

The completion scrips complete the second argument of the command. And
short instructions on how to test the completion updates are added to
README.md.