Add `show-workflows` command
The command aims to show all available personal and common workflows
for the repository.

Adding of the command encourages to change the layout of the main help
message. The updated help groups the commands based on their purposes
rather than by the type of objects they interact with.

Also, for the optimization of further development, the `workflows`
plugin is introduced. It serves all workflows-specific common logic.