Read multi slash branch names when `deliver-work`
`git elegant deliver-work` incorrectly reads the name of the upstream
branch if it contains at least one slash like `feature/123`. This is
because the used pattern matching logic doesn't expect remote branch
names like `origin/feature/123`.

In order to solve it, the new `transformation` plugin is introduced. It
provides functions that parse remote branch names correctly. Now, the
plugin handles branch translation (mapping) for `deliver-work` command
as well as for `accept-work` and `obtain-work` in order to prevent
possible errors in the future.

#255