Standardize wording across all documents
First of all, all used terminology is aligned with Git glossary (`git
help gitglossary`). It's also relevant for the names of some variables
in the scripts.

Next change is the update of the documentation of the command for both command-line
help and WEB documentation. The update tries to express the meaning and
logic of the commands in an unambiguous manner.

All trailing whitespaces are removed from all project files.

Last but not least, don't add `bash` syntax for describing `git-elegant`
help during the `commands.md` generation  as it is just a message.