Print error with a relevant command for docs generation
The commit fixes the error message that is displayed if the docs weren't
updated prior to making a commit. The problem is that the command for
documentation generation is wrong. Now it's updated to use
`.workflows/docs.py` script instead of the old Bash script (which was
removed some time ago).