Add logic of protecting branches
A user may have several branches in the repository in which he/she can't
do modifications. For instance, it is a case when there is `master` and
`develop` branches. These branches are protected ones and Elegant Git
allows to:
- configure global and local branches that should be protected
- disable removing of the protected branches
- disable changing the history of protected branches

This is implemented by
1. aborting the execution of `amend-work`, `deliver-work`, `polish-work`,
`save-work` if any of the commands is run against a protected branch
2. skipping the protected branches in `prune-repository` command
3. asking for the protected branches in the `acquire-git` and
   `acquire-repository` commands

Also, the documentation is updated accordingly.