Don't set protected branches in global configuration
By default, the `master` branch is protected even if the protected
branches were not configured. So, if a user doesn't run any
configuration, the `master` branch will be protected. If a user wants
to protect some other branch in the current (some) repository, then the
`acquire-repository` has to be executed - sets the protected branches.
So, there is no sense to configure protected branches during the global
configuration (`acquire-git`).