Make the default development branch configurable
Up to now, the `master` branch was the default development branch. And
now it's possible to configure it by running the `acquire-repository`
command. So, after the configuration, the provided value will be used
instead of the `master`. However, the `master` is still a default one.
But now, it's possible to set another branch as a default one (for
instance, `develop`) and Elegant Git will use it.