Use `bats-core` instead of `bats`
There is a separate Docker worker that allows running Bats tests based
on version 1.2.0 of `bats-core` project. This modification allows
receiving updates to the test runner.

All testing procedure is reworked in a way that `workflows` file is
responsible for running Docker workers for testing while scripts,
located at `.workflows` directory, just run the needed workflows without
pointing to an environment.

`tests/git-elegant.bats` was updated as in Bash 3.x (which is used in
the Bats worker) is not possible to get the last element of the array.
So, `show-commands` in the tests is replaced with `prune-repository` as
the last command produces the same amount of lines comparing to the fact
that adding or removing of command will affect the output of
`show-commands`.

#135