Use the latest version of the default branch commits
There is a local repository that has an associated remote. If Elegant
Git is used for serving interactions with it, it may happen that
the default local branch won't have all commits that are presented in
the local remote branch. This happens because Elegant Git uses the
`fetch` operation to get the latest changes that update the local
remote branches, not the local branches. This leads to not up-to-date
outputs of commands such as `polish-work`, `show-work`, etc.

From the implementation side, the default branch (`master` or any other
configured) is used as the source of the latest commits. However,
`origin/master` can have more commits then `master`.

That's why `origin/master` is enforced to be used in such cases. This
guarantee that Elegant Git will use the latest available commits.