Use help message of desired command
Disable backtracking in cli parser.

Explanation:
In haskell, we have a pattern which allows us to try multiple options
untill something succeedes. If option failed, we backtrack to the parent
to try more. This has an advantage of allowing a high-level code to be
insanely flexible.
This is where a problem lies: if we find the command `show-work` but
then fail to parse arguments, with the `simple` implementation we have
no way to know that we should commit to parse this message. In our case,
the ideal approach is to just disable backtracking, as this will force
parser to try parsing a specific command as soon as it found one.

#289