Improve testability of the `Real` interpreter
There is an issue in how we test actions. Fundamentally, we rely on the
fact that the `Real` runner is generating `git` calls correctly. This
could be neglected in theory, but there is another issue, which forces
us to generate "textual" representation of what our command is doing, in
multiple places, which leads to code repetition.
As a solution to this, I've created 1 data type for every type of the
command we can execute, and implemeted the way to render it to the
"textual" representation.
This could be used by `Real` runner directly, but this still suffers
from a different issue. We give `Real` runner to much access, by adding
`MonadIO` to the context.
There is a solution however. We can create a custom `Monad` which can
provide all needed functions for `Real` to proceed. Going this way, we
can hide how we "render" and "execute" git calls from the `Real` runner,
which implies that it has no way to run arbitrary commands at all.

#299