Implement `type` command to get gpg path
`type` command is only available as the shell command which means that
we can't use `proc`. This means we should use `shell` instead.
However, using `shell` is not safe, and escaping is needed to make sure
args are mungled. This is the reason why I use `shell` only for `type`
command and `proc` for everything else.

#297